% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{api_cost}
\alias{api_cost}
\title{Estimate API cost}
\usage{
api_cost(prompt_tokens, completion_tokens, selected_model)
}
\arguments{
\item{prompt_tokens}{a number}

\item{completion_tokens}{a number}

\item{selected_model}{a string}
}
\value{
a number
}
\description{
Estimate API cost
}
