% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hints.r
\name{auto.hint}
\alias{auto.hint}
\title{This is just a place holder in a hint block}
\usage{
auto.hint()
}
\description{
Only used inside a hint block.
}
\details{
It says that the automatic hint shall be shown. This
makes sense if you want to show the automatic hint in
addition to a custom hint. Also see auto_hint_else()
}
