% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_for_ps.r
\name{check.assign.with.multiple.sol}
\alias{check.assign.with.multiple.sol}
\title{Checks an assignment to a variable with up to 5
possibly correct solutions}
\usage{
check.assign.with.multiple.sol(sol1, sol2, sol3, sol4, sol5, ...,
  sol.list = list())
}
\arguments{
\item{sol1}{An assignment that needs to be checked, e.g. x<-5. Similar for sol2, sol3, sol4, sol5.}
}
\description{
Can be called in a #< test block for a custom test.
}
\examples{
# Assume the task is that x shall be a number 
# below 11 and divisible by 5

check.assign.with.multiple.sol(x<-5, x<-10)
}
