% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_expr.r
\name{check.call.args}
\alias{check.call.args}
\title{Checks whether arguments of stud.call are correct given the specification in check.call}
\usage{
check.call.args(stud.call, check.call, compare.vals = !is.null(val.env),
  val.env = NULL, allow.extra.arg = FALSE, ignore.arg = NULL,
  check.values = NULL)
}
\description{
Checks whether arguments of stud.call are correct given the specification in check.call
}

