% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sol.r
\name{check.chunk}
\alias{check.chunk}
\title{Check the student's solution of an exercise contained in stud.file}
\usage{
check.chunk(chunk.ind, ps = get.ps(), verbose = FALSE,
  stud.code = ps$cdt$stud.code[[chunk.ind]],
  stud.env = make.chunk.stud.env(chunk.ind, ps), expect.change = FALSE,
  store.output = TRUE, noeval = isTRUE(ps$noeval),
  precomp = isTRUE(ps$precomp))
}
\arguments{
\item{stud.code}{The code of the student's solution as a string (or vector of strings)}

\item{ex.name}{The name of the exercise}
}
\description{
Check the student's solution of an exercise contained in stud.file
}

