\name{check.H0.rejected}
\alias{check.H0.rejected}
\title{Check whether a certain H0 could be significantly rejected with a specified statistical test}
\usage{
  check.H0.rejected(test.expr, alpha = 0.05,
    failure.message, student.env = get.student.env())
}
\description{
  Check whether a certain H0 could be significantly
  rejected with a specified statistical test
}

