% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_for_ps.r
\name{check.package}
\alias{check.package}
\title{Tests whether a package is loaded. Not yet implemented for speed reasons}
\usage{
check.package(package)
}
\description{
Tests whether a package is loaded. Not yet implemented for speed reasons
}

