% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check.problem.set}
\alias{check.problem.set}
\title{Checks a student problem set}
\usage{
check.problem.set(ps.name, stud.path, stud.short.file, reset = FALSE,
  set.warning.1 = TRUE, user.name = "GUEST", do.check = interactive(),
  verbose = FALSE, catch.errors = TRUE)
}
\description{
The command will be put at the top of a student's problem set. It checks all exercises when the problem set is sourced. If something is wrong, an error is thrown and no more commands will be sourced.
}

