\name{check.var}
\alias{check.var}
\title{Test: Compare students variables with those from the given solutions}
\usage{
  check.var(vars, exists = FALSE, length = FALSE,
    class = FALSE, values = FALSE, verbose = FALSE,
    ex = get.ex(), stud.env = ex$stud.env,
    sol.env = ex$sol.env)
}
\description{
  Test: Compare students variables with those from the
  given solutions
}

