% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{clean_api_key}
\alias{clean_api_key}
\title{Clean up API key character}
\usage{
clean_api_key(api_key)
}
\arguments{
\item{api_key}{is a character string}
}
\value{
Returns a string with api key.
}
\description{
The response from GPT3 sometimes contains strings that are not R commands.
}
