% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{clean_code}
\alias{clean_code}
\title{Clean up text strings for inserting into SQL}
\usage{
clean_code(x)
}
\arguments{
\item{x}{a string that can contain ' or "}
}
\value{
nothing
}
\description{
Clean up text strings for inserting into SQL
}
