% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_expr.r
\name{compare.calls}
\alias{compare.calls}
\title{Compare if two calls are the same}
\usage{
compare.calls(stud.call, check.call, compare.vals = !is.null(val.env),
  val.env = NULL, ...)
}
\description{
Compare if two calls are the same
}

