% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_ps.r
\name{copy.ps.for.session}
\alias{copy.ps.for.session}
\title{Makes a local copy of a problem set for a new shiny session}
\usage{
copy.ps.for.session(ps, empty.stud.env = TRUE)
}
\description{
Makes a local copy of a problem set for a new shiny session
}

