\name{create.empty.ps.and.rps}
\alias{create.empty.ps.and.rps}
\title{Generate an empty problem set for a student in folder dir and the rps file}
\usage{
create.empty.ps.and.rps(ps.name = get.ps()$name, dir = getwd(),
  user.name = "ENTER A USER NAME HERE", header = "", footer = "",
  libs = NULL, make.rps = TRUE)
}
\description{
Generate an empty problem set for a student in folder dir
and the rps file
}

