\name{create.empty.ps}
\alias{create.empty.ps}
\title{Generate an problem set for a student in folder dir}
\usage{
  create.empty.ps(ps.name = get.ps()$name, dir = getwd(),
    user.name = "ENTER A USER NAME HERE", header = "",
    footer = "")
}
\description{
  Generate an problem set for a student in folder dir
}

