\name{create.ps}
\alias{create.ps}
\title{Creates a problem set structure files and empty .r and .rmd solution files}
\usage{
create.ps(sol.file, ps.name, ...)
}
\arguments{
  \item{sol.file}{file name of the solution file}

  \item{ps.name}{name of the problem set}
}
\description{
Creates a problem set structure files and empty .r and .rmd
solution files
}

