\name{create.sample.solution.from.rmd}
\alias{create.sample.solution.from.rmd}
\title{Creates a problem set structure file from a solution file}
\usage{
create.sample.solution.from.rmd(sol.file, target.file = NULL,
  ps.name = NULL, user.name = "Jane Doe", dir = getwd(), libs = NULL,
  header = "", footer = "")
}
\arguments{
  \item{sol.file}{file name of the solution file}

  \item{ps.name}{name of the problem set}
}
\description{
Creates a problem set structure file from a solution file
}

