\name{create.struc.from.rmd}
\alias{create.struc.from.rmd}
\title{Creates a problem set structure file from a .rmd solution file}
\usage{
create.struc.from.rmd(sol.file, ps.name = NULL)
}
\arguments{
  \item{sol.file}{file name of the solution file}

  \item{ps.name}{name of the problem set}
}
\description{
Creates a problem set structure file from a .rmd solution
file
}

