\name{create.struc}
\alias{create.struc}
\title{Creates a problem set structure file from a solution file}
\usage{
create.struc(sol.file, ps.name = NULL)
}
\arguments{
  \item{sol.file}{file name of the solution file}

  \item{ps.name}{name of the problem set}
}
\description{
Creates a problem set structure file from a solution file
}

