\name{create.stud.ps}
\alias{create.stud.ps}
\title{Generate a problem set skeleton for a student and save it in a file}
\usage{
  create.stud.ps(ps, file = ps$stud.file,
    ps.dir = "C:/...", game.mode = FALSE)
}
\description{
  Generate a problem set skeleton for a student and save it
  in a file
}

