% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_usage_db}
\alias{create_usage_db}
\title{Creates a SQLite database file for collecting user data}
\usage{
create_usage_db()
}
\value{
nothing
}
\description{
The data file should be stored in the ../../data folder inside 
the container. From outside in the RTutor_server folder, 
it is in data folder.
 Only works on local machines. Not on linux.
}
