% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_rtutor.r
\name{data}
\alias{data}
\title{Overwrite the base function data, copy data by default into the calling environment instead of the global environment}
\usage{
data(..., envir = parent.frame())
}
\description{
Overwrite the base function data, copy data by default into the calling environment instead of the global environment
}

