% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ups.r
\name{default.ups.save}
\alias{default.ups.save}
\title{Specify which information will be automatically saved in ups}
\usage{
default.ups.save(chunks = TRUE, awards = TRUE, addons = TRUE,
  code = FALSE | all, chunk.ind = FALSE | all, all = FALSE)
}
\description{
Specify which information will be automatically saved in ups
}
