% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{describe_data}
\alias{describe_data}
\title{Generate a plain text about the distribution and correlations.}
\usage{
describe_data(df)
}
\arguments{
\item{df}{a data frame.}
}
\value{
a text string
}
\description{
Generate a plain text about the distribution and correlations.
}
