\name{develop.problem.set}
\alias{develop.problem.set}
\title{Init files for developing a new problem set with given name under the specified parent directory}
\usage{
  develop.problem.set(name, parent.path)
}
\description{
  Init files for developing a new problem set with given
  name under the specified parent directory
}

