% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{distribution_description}
\alias{distribution_description}
\title{Describes the distribution of a column in a data frame, or a vector.}
\usage{
distribution_description(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a text string
}
\description{
Describes the distribution of a column in a data frame, or a vector.
}
