% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{estimate_tokens}
\alias{estimate_tokens}
\title{Estimate tokens from text}
\usage{
estimate_tokens(text)
}
\arguments{
\item{text}{a string}
}
\value{
a number
}
\description{
Estimate tokens from text
}
