% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sol.r
\name{extract.exercise.code}
\alias{extract.exercise.code}
\title{Extracts the stud's code of a given exercise}
\usage{
extract.exercise.code(ex.name, stud.code = ps$stud.code, ps = get.ps(),
  warn.if.missing = TRUE)
}
\description{
Extracts the stud's code of a given exercise
}

