% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ps.r
\name{extract.ps.title}
\alias{extract.ps.title}
\title{Extracts problem set title}
\usage{
extract.ps.title(txt, te.name)
}
\arguments{
\item{txt}{RMD file of problemset}

\item{te.name}{value which is used as backup, probably te$ps.name}
}
\description{
Returns either the first top-level header (prefered) or the value given by te.name. The top-level header has to be before all other ones.
}
