% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{extract_calls}
\alias{extract_calls}
\title{Recursively extract every function name called in a parsed R expression}
\usage{
extract_calls(parsed_exprs)
}
\arguments{
\item{parsed_exprs}{result of parse(text = ...)}
}
\value{
character vector of unique called function names
}
\description{
Handles: plain calls foo(), qualified calls pkg::foo(),
do.call("foo", ...) and get("foo")() with string literals.
}
