% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\docType{data}
\name{faqs}
\alias{faqs}
\title{Saves user queries, code, and error status}
\format{
An object of class \code{data.frame} with 9 rows and 2 columns.
}
\usage{
faqs
}
\arguments{
\item{request, }{user request}

\item{code}{AI generated code}

\item{error}{status, TRUE, error}

\item{chunk, }{id, from 1, 2, ...}

\item{api_time}{time in seconds for API response}

\item{tokens}{total completion tokens}

\item{filename}{name of the uploaded file}

\item{filesize}{size}

\item{date}{Date in the format of "2023-01-04"}

\item{time}{Time "13:05:12"}

\item{comments, }{user request}

\item{helpfulness}{rating}

\item{experience}{R experience}
}
\value{
nothing
Save user feedback

nothing
}
\description{
Saves user queries, code, and error status
}
\keyword{datasets}
