% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeUtils.r
\name{get.current.file.path}
\alias{get.current.file.path}
\title{Returns file path of current file}
\usage{
get.current.file.path()
}
\value{
A character value defining the current path.
}
\description{
When called out of RStudio API returns the current path of the active file without the file name.
}
\examples{
setwd(get.current.file.path())

}
