% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTutor.R
\name{get.ps}
\alias{get.ps}
\title{Get the current problem set}
\usage{
get.ps(force.global = FALSE)
}
\description{
Either a globally stored problem set or
if RTutor runs as a web-app the associated problem set
with the current shiny session
}

