\name{give.award}
\alias{give.award}
\title{Used in a solution file. Give an award to a student who made it so far}
\usage{
give.award(award, text = "", level = NULL, prize.text = NULL,
  prize.code = NULL, user = get.user())
}
\description{
Used in a solution file. Give an award to a student who
made it so far
}

