% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hints.r
\name{hint.else.active}
\alias{hint.else.active}
\title{Get or set whether hint.else or
auto.hint.else would be triggered.}
\usage{
hint.else.active(activate = NULL, ps = get.ps())
}
\description{
If a hint.stud.call or hint.stud.assign is shown
then a hint.else or auto.hint.else would not be
triggered. This function returns TRUE if hint.else
would still be triggered or otherwise FALSE.
}
\details{
If you set the argument activate you can change this status.
}
