% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hints.r
\name{hint.else}
\alias{hint.else}
\title{Show a hint only if no hint.stud.call or hint.stud.assign 
was triggered.}
\usage{
hint.else(msg, add.line.breaks = TRUE, ps = get.ps())
}
\description{
It says that the automatic hint should be shown unless
some hint with hint.stud.call has been shown (or ps$shown.custom.hints has been manually assigned a value above 0.)
}
