% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hints.r
\name{hint.for.assign}
\alias{hint.for.assign}
\title{Default hint for an assignment}
\usage{
hint.for.assign(expr, ps = get.ps(), env = ps$stud.env,
  stud.expr.li = ps$stud.expr.li, part = ps$part, s3.method = NULL,
  expr.object = NULL, start.char = "\\n", end.char = "\\n", ...)
}
\description{
Default hint for an assignment
}

