% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hints.r
\name{hint.for.compute}
\alias{hint.for.compute}
\title{Default hint for a compute block}
\usage{
hint.for.compute(expr, hints.txt = NULL, var = "", ps = get.ps(),
  env = ps$stud.env, stud.expr.li = ps$stud.expr.li, part = ps$part, ...)
}
\description{
Default hint for a compute block
}

