% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hints.r
\name{hint.stud.assign}
\alias{hint.stud.assign}
\title{Show the hint if the student made the specified wrong assignment}
\usage{
hint.stud.assign(var, call, msg, ps = get.ps(), env = parent.frame())
}
\arguments{
\item{var}{name of the to be assigned variable as character}

\item{call}{an unquoted call that we check whether the student makes it}

\item{msg}{a string that shall be shown as hint if the student made the call in his code}
}
\description{
Show the hint if the student made the specified wrong assignment
}
