\name{holds.true}
\alias{holds.true}
\title{Test: Does a certain condition on the user's generated variables hold true}
\usage{
  holds.true(cond,
    failure.message = "Failure in holds.true",
    student.env = get.student.env())
}
\description{
  Test: Does a certain condition on the user's generated
  variables hold true
}

