% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_ps.r
\name{init.ps}
\alias{init.ps}
\title{Initialize a problem set for the student}
\usage{
init.ps(ps.name, dir = getwd(), stud.short.file = paste0(ps.name, ".Rmd"),
  rps.file = paste0(rps.dir, "/", ps.name, ".rps"), log.file = paste0(dir,
  "/", ps.name, ".log"), rps.dir = dir, save.nothing = FALSE,
  check.whitelist = !is.null(wl), wl = NULL, use.memoise = NULL,
  noeval = FALSE, precomp = FALSE)
}
\arguments{
\item{ps.name}{the name of the problem set}

\item{dir}{the path in which the stud has stored his file}

\item{stud.hort.file}{the file name (without path) of the .rmd problem set file}
}
\description{
Initialize a problem set for the student
}

