% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{input_search}
\alias{input_search}
\title{Search User input.}
\usage{
input_search(prepared_request)
}
\arguments{
\item{prepared_request}{A string that stores the user input.}
}
\value{
Returns specified instructions to be sent to GPT.
}
\description{
Add additional info to prompt for certain user inputs.
}
