% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_plotter.R
\name{launch_plotter}
\alias{launch_plotter}
\title{Launch Plotter}
\usage{
launch_plotter(list_of_datasets)
}
\arguments{
\item{list_of_datasets}{List of datasets which should be made available for selection when the app is launched}
}
\value{
Launches App
}
\description{
Launches the plotting app
}
\examples{
\dontrun{
diamonds_sample <- ggplot2::diamonds[sample(1:nrow(diamonds),size=1000),]
launch_plotter(list(diamonds_sample = diamonds_sample, mtcars = mtcars, iris = iris))}
}
