% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tut_code.R
\name{launch_tutorial}
\alias{launch_tutorial}
\title{Launch Tutorial App}
\usage{
launch_tutorial()
}
\value{
Launches app
}
\description{
Invoking this function will launch the Tutorial App;
The app includes a ReadMe introduction which provides a quick overview on how to use the app
}
\examples{
# Simply invoke the function to launch the app
\dontrun{
launch_tutorial()}
}

