% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_ps.r
\name{list.ps}
\alias{list.ps}
\title{Returns a list of the names of all problem sets that are included in RTutor}
\usage{
list.ps()
}
\description{
Returns a list of the names of all problem sets that are included in RTutor
}

