\name{load.problem.set}
\alias{load.problem.set}
\title{Load the structure of a problem set}
\usage{
  load.problem.set(name,
    file = paste0(get.problemset.structure.path(), "ps_", name, "_struc.r"))
}
\arguments{
  \item{name}{the name of the problem set}

  \item{the}{r file in which the structure of the problem
  set is described}
}
\description{
  Load the structure of a problem set
}

