% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hint_report.r
\name{make.hint.report}
\alias{make.hint.report}
\title{Helper function when developing problem sets}
\usage{
make.hint.report(ps.name, out.file = paste0(ps.name, "_hint_report.Rmd"))
}
\description{
Tries to check all chunks with given solution and shows
the corresponding hint.
}
