% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{move_front}
\alias{move_front}
\title{Move an element to the front of a vector}
\usage{
move_front(v, e)
}
\arguments{
\item{v}{is the vector}

\item{e}{is the element}
}
\value{
Returns a reordered vector
}
\description{
The response from GPT3 sometimes contains strings that are not R commands.
}
