% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ps.r
\name{name.rmd.chunks}
\alias{name.rmd.chunks}
\title{Set default names for the chunks of problem set rmd files}
\usage{
name.rmd.chunks(rmd.file = NULL, txt = readLines(rmd.file),
  only.empty.chunks = FALSE, keep.options = TRUE, valid.file.name = FALSE)
}
\arguments{
\item{rmd.file}{file name}

\item{txt}{alternative the code as txt file}

\item{only.empy.chunks}{if FALSE (default) name all chunks.
Otherwise only empty chunks are overwritten}

\item{keep.option}{if TRUE (default) don't change chunk options;
otherwise clear all chunk options (dangerous)}
}
\description{
Set default names for the chunks of problem set rmd files
}

