% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{numeric_to_factor}
\alias{numeric_to_factor}
\title{Returns a data frame with some numeric columns with fewer levels 
converted as factors}
\usage{
numeric_to_factor(df, max_levels_factor, max_proportion_factor)
}
\arguments{
\item{df}{a data frame}

\item{max_levels_factor}{max levels, defaults to max_levels}

\item{max_proportion_factor}{max proportion}
}
\value{
Returns a data frame
}
\description{
Returns a data frame with some numeric columns with fewer levels 
converted as factors
}
