% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_frame_ps.r
\name{parse.chunk.args}
\alias{parse.chunk.args}
\title{Parse the name of a knitr chunk and its arguments}
\usage{
parse.chunk.args(header, arg.str = NULL)
}
\description{
Parse the name of a knitr chunk and its arguments
}
