% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{polish_cmd}
\alias{polish_cmd}
\title{Remove Markdown and explanation}
\usage{
polish_cmd(cmd)
}
\arguments{
\item{cmd}{A string that stores the completion from GTP3.}
}
\value{
Returns a cleaned up version, so that it could be executed as R command.
}
\description{
The response from GTP3 sometimes contains strings that are not R commands.
}
