% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{python_html}
\alias{python_html}
\title{Generate html file from Python code}
\usage{
python_html(python_code, select_data, current_data)
}
\arguments{
\item{python_code, }{a chunk of code}

\item{select_data}{input$select data}

\item{current_data}{current_data()}

\item{html_file}{file name for output}
}
\value{
-1 if failed. If success, the the designated html file is written
}
\description{
Generate html file from Python code
}
