% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{read_additional_data}
\alias{read_additional_data}
\title{Read built-in datasets}
\usage{
read_additional_data()
}
\value{
Returns a cleaned up version, so that it could be sent to GPT.
}
\description{
Read user data files
}
