% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{renderCode}
\alias{renderCode}
\alias{codeOutput}
\title{Render code with syntax highlighting}
\usage{
renderCode(
  expr,
  env = parent.frame(),
  quoted = FALSE,
  outputArgs = list(),
  delay = 100
)

codeOutput(outputId)
}
\arguments{
\item{expr}{An expression to evaluate.}

\item{env}{The parent environment for the reactive expression. By default,
this is the calling environment, the same as when defining an ordinary
non-reactive expression. If \code{expr} is a quosure and \code{quoted} is \code{TRUE},
then \code{env} is ignored.}

\item{quoted}{If it is \code{TRUE}, then the \code{\link[=quote]{quote()}}ed value of \code{expr}
will be used when \code{expr} is evaluated. If \code{expr} is a quosure and you
would like to use its expression as a value for \code{expr}, then you must set
\code{quoted} to \code{TRUE}.}

\item{outputArgs}{A list of arguments to be passed through to the implicit
call to [codeOutput] when `renderCode` is used in an interactive R
Markdown document.}

\item{delay}{How much delay to use (in ms) between rendering and firing the
syntax highlighting script. For `length(delay) > 1` the script will fire
multiple times.}

\item{outputId}{output variable to read the code from.}
}
\description{
These functions work similar to `renderText` and `verbatimTextOutput` but
will include highlighting with `highlight.js` and `rstudio.css`.
}
