% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.r
\name{rtutor.observe.html.query}
\alias{rtutor.observe.html.query}
\title{This function must be called in the initHandler of the app}
\usage{
rtutor.observe.html.query(app = getApp(), ps = get.ps(),
  session.timeout.sec = 300)
}
\description{
This function must be called in the initHandler of the app
}
