% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_package.r
\name{rtutor.skel.show.opts.string}
\alias{rtutor.skel.show.opts.string}
\title{Intermediary Function helping to build the ps.show() Options string}
\usage{
rtutor.skel.show.opts.string(mandatory, optional)
}
\arguments{
\item{mandatory}{Are always set but may be overwritten by optional}

\item{optional}{Are intended to be set by the user. May overwrite mandatory ones if set explicitely.}
}
\description{
Expects two lists with arguments.
}
