% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{save_comments}
\alias{save_comments}
\title{Saves user queries, code, and error status}
\usage{
save_comments(date, time, comments, helpfulness, experience)
}
\arguments{
\item{date}{Date in the format of "2023-01-04"}

\item{time}{Time "13:05:12"}

\item{comments, }{user request}

\item{helpfulness}{rating}

\item{experience}{R experience}

\item{request, }{user request}

\item{code}{AI generated code}

\item{error}{status, TRUE, error}

\item{chunk, }{id, from 1, 2, ...}

\item{api_time}{time in seconds for API response}

\item{tokens}{total completion tokens}

\item{filename}{name of the uploaded file}

\item{filesize}{size}
}
\value{
nothing
Save user feedback

nothing
}
\description{
Saves user queries, code, and error status
}
