% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{save_data}
\alias{save_data}
\title{Saves user queries, code, and error status}
\usage{
save_data(date, time, request, code, error_status, data_str, dataset)
}
\arguments{
\item{date}{Date in the format of "2023-01-04"}

\item{time}{Time "13:05:12"}

\item{request, }{user request}

\item{code}{AI generated code}

\item{error}{status, TRUE, error}
}
\value{
nothing
}
\description{
Saves user queries, code, and error status
}
