% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_rtutor.r
\name{sc}
\alias{sc}
\title{Like paste0 but returns an empty vector if some string is empty}
\usage{
sc(..., sep = "", collapse = NULL)
}
\description{
Like paste0 but returns an empty vector if some string is empty
}

