% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeUtils.r
\name{set.wd.current}
\alias{set.wd.current}
\title{Sets Working directory to current directory}
\usage{
set.wd.current()
}
\value{
-
}
\description{
When called out of RStudio API or per Source sets the working directory to the current directory. If called from console nothing happens.
}
