% Generated by roxygen2 (4.0.1): do not edit by hand
\name{show.shiny.ps}
\alias{show.shiny.ps}
\title{Run a problem set in the webbroser (or in the viewer pane).}
\usage{
show.shiny.ps(ps.name, user.name = "Seb", sav.file = NULL,
  load.sav = !is.null(sav.file), sample.solution = FALSE,
  is.solved = load.sav, import.rmd = FALSE, rmd.file = paste0(ps.name,
  "_", user.name, "_export.rmd"), launch.browser = TRUE,
  catch.errors = TRUE, ...)
}
\arguments{
\item{load.sav}{shall the last saved be loaded?}

\item{sample.solution}{shall the sample solution be shown}

\item{is.solved}{if sample.solution or load.sav shall the correct chunks be automatically run when the problem set is loaded?}

\item{import.rmd}{shall the solution be imported from the rmd file specificed in the argument rmd.file}

\item{lauch.browser}{if TRUE (default) show the problem set in the browser. Otherwise it is shown in the RStudio viewer pane}

\item{catch.errors}{by default TRUE only set FALSE for debugging purposes in order to get a more informative traceback()}
}
\description{
... contains parameters specified in init.shiny.ps. They are explained here.
}

