% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_for_ps.r
\name{show.success.message}
\alias{show.success.message}
\title{Simply shows a success message when this test is reached for the first time!}
\usage{
show.success.message(success.message, ...)
}
\description{
Simply shows a success message when this test is reached for the first time!
}

