% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tut_code.R
\name{show_tutorial_datasets}
\alias{show_tutorial_datasets}
\title{Extract Tutorial datasets}
\usage{
show_tutorial_datasets()
}
\value{
A list object containing the tutorial datasets for each topic
}
\description{
Extract the datasets used by the tutorial app
}
\examples{
# If you would like to extract the underlying datasets used for building the app
# you can invoke this function
show_tutorial_datasets()
}

