% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_for_ps.r
\name{sigma.test}
\alias{sigma.test}
\title{Test for variance (copied from TeachingDemos)}
\usage{
sigma.test(x, sigma = 1, sigmasq = sigma^2, alternative = c("two.sided",
  "less", "greater"), conf.level = 0.95, ...)
}
\description{
Test for variance (copied from TeachingDemos)
}

