% Generated by roxygen2 (4.0.1): do not edit by hand
\name{test.H0.rejected}
\alias{test.H0.rejected}
\title{Test whether a certain H0 can be significantly rejected}
\usage{
test.H0.rejected(test.expr, p.value, test.name = "", alpha.warning = 0.01,
  alpha.failure = 0.05,
  short.message = "Fail to reject '{{test_name}}', p.value = {{p_value}}",
  warning.message = "The null hypothesis from the test '{{test_name}}', should not be rejcected, but I get a fairly low p.value of {{p_value}}.",
  failure.message = "I couldn't significantly reject the null hypothesis from the test '{{test_name}}', p.value = {{p_value}}",
  success.message = "Great, I could significantly reject the null hypothesis from the test '{{test_name}}', p.value = {{p_value}}!",
  check.warning = TRUE, ps = get.ps(), stud.env = ps$stud.env,
  part = NULL, ...)
}
\arguments{
\item{test.expr}{an expression that calls a test which will be evaluated in stud.env. The test must return a list that contains a field "p.value"}

\item{p.value}{Instead of providing test.expr, one can directly provide a p.value from a previously run test}

\item{test.name}{an optional test.name that can be used to fill the {{test_name}} whiskers in warning or failure messages.}

\item{alpha.failure}{default=0.001 the critical p.value below which the stud code is considered wrong}

\item{alpha.warning}{default=0.05 a p.value below a warning is printed that the code may be wrong}

\item{short.message,failure.messages,}{warning.messages Messages in case of a failure and warning and  short message for the log.file}

\item{check.warning}{if FALSE don't check for a warning}
}
\value{
TRUE if H0 can be rejected, FALSE if not and "warning" if it can be weakly rejected
}
\description{
Test whether a certain H0 can be significantly rejected
}

