\name{test.normality}
\alias{test.normality}
\title{Test: Has a vector of random numbers been drawn from a normal distribution?}
\usage{
  test.normality(vec, short.message, warning.message,
    failure.message, ex = get.ex(), stud.env = ex$stud.env,
    ...)
}
\description{
  Test: Has a vector of random numbers been drawn from a
  normal distribution?
}

