\name{test.variance}
\alias{test.variance}
\title{Test: The variance of the distribution from which a vector of random numbers has been drawn}
\usage{
  test.variance(vec, true.val, alpha.warning = 0.05,
    alpha.fail = 1e-04, test = "t.test",
    student.env = get.student.env(), ..., digits = 7,
    verbose = FALSE)
}
\description{
  Test: The variance of the distribution from which a
  vector of random numbers has been drawn
}

