% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hints.r
\name{true}
\alias{true}
\title{A robust implementation of isTRUE}
\usage{
true(expr, envir = parent.frame())
}
\description{
Returns FALSE if evaluation expr yields an
error or is not TRUE.
}
\details{
Useful for customized hints were evaluating
an expression may often cause errors,
e.g. if a user did not define a variable.
}
