% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{turned_on}
\alias{turned_on}
\title{Returns true only defined and has a value of true}
\usage{
turned_on(x)
}
\arguments{
\item{x}{}
}
\value{
Returns TRUE or FALSE
}
\description{
This is used when some the input variables are not defined globally,
But could be turned on. if you use if(input$selected), it will
give an error.
}
