% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\docType{data}
\name{user_upload}
\alias{user_upload}
\title{No Data' # no data is uploaded or selected}
\format{
An object of class \code{character} of length 1.
}
\usage{
user_upload
}
\description{
No Data' # no data is uploaded or selected
}
\keyword{datasets}
