% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{validate_api_key}
\alias{validate_api_key}
\title{Validate API key character}
\usage{
validate_api_key(api_key)
}
\arguments{
\item{api_key}{is a character string}
}
\value{
Returns TRUE or FALSE
}
\description{
The response from GPT3 sometimes contains strings that are not R commands.
}
