% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_rtutor.r
\name{with.random.seed}
\alias{with.random.seed}
\title{Calls a function with a specified random.seed}
\usage{
\method{with}{random.seed}(expr, seed = 1234567890)
}
\arguments{
\item{expr}{the expression to be evaluated}

\item{seed}{the seed as integer}
}
\description{
Calls a function with a specified random.seed
}
