\name{write.ps.skeleton}
\alias{write.ps.skeleton}
\title{Generate a problem set skeleton for a student and save it in a file}
\usage{
  write.ps.skeleton(ps,
    file = paste0(get.problemset.user.path(), "ps_", ps$name, "_skeleton.r"))
}
\description{
  Generate a problem set skeleton for a student and save it
  in a file
}

