\name{zip.solution}
\alias{zip.solution}
\title{Create a zip file of a students solution that can be uploaded}
\usage{
  zip.solution(ps.name = get.ps()$name,
    user.name = get.user()$name, dir = get.ps()$stud.path,
    make.html = TRUE)
}
\description{
  Create a zip file of a students solution that can be
  uploaded
}

