% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_ps.r
\name{zip.submit.footer.txt}
\alias{zip.submit.footer.txt}
\title{Generate default footer text for a Rmd file}
\usage{
zip.submit.footer.txt(ps.name)
}
\description{
Generate default footer text for a Rmd file
}

