% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add.success}
\alias{add.success}
\title{Used inside tests: adds a sucess message}
\usage{
add.success(message, ..., ps = get.ps())
}
\arguments{
\item{message}{a longer description shown to the user}

\item{...}{variables that will be rendered into messages that have whiskers}
}
\description{
Used inside tests: adds a sucess message
}

