% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sol.r
\name{add.warning}
\alias{add.warning}
\title{Used inside tests: adds a warning}
\usage{
add.warning(message, ..., ps = get.ps())
}
\arguments{
\item{message}{a longer description shown to the user}

\item{...}{variables that will be rendered into messages that have whiskers}
}
\description{
Used inside tests: adds a warning
}

