% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hints.r
\name{auto.hint.else}
\alias{auto.hint.else}
\title{This is just a place holder in a hint block}
\usage{
auto.hint.else()
}
\description{
It says that the automatic hint should be shown unless
some hint with hint.stud.call has been shown (or ps$shown.custom.hints has been manually assigned a value above 0.)
}
