% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check.assign}
\alias{check.assign}
\title{Checks an assignment to a variable}
\usage{
check.assign(call, check.arg.by.value = TRUE, allow.extra.arg = FALSE,
  ignore.arg = NULL, success.message = NULL, failure.message = NULL,
  no.command.failure.message = "You have not yet included correctly, all required R commands in your code...",
  ok.if.same.val = TRUE, call.object = NULL, s3.method = NULL,
  ps = get.ps(), stud.env = ps$stud.env, part = ps$part,
  stud.expr.li = ps$stud.expr.li, verbose = FALSE, ...)
}
\description{
Checks an assignment to a variable
}

