% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check.call}
\alias{check.call}
\title{Checks whether the user makes a particular function call in his code or call a particular R statement}
\usage{
check.call(call, check.arg.by.value = TRUE, allow.extra.arg = FALSE,
  ignore.arg = NULL, success.message = NULL, failure.message = NULL,
  no.command.failure.message = NULL, ok.if.same.val = FALSE,
  s3.method = NULL, ps = get.ps(), stud.env = ps$stud.env,
  part = ps$part, stud.expr.li = ps$stud.expr.li, verbose = FALSE, ...)
}
\description{
Checks whether the user makes a particular function call in his code or call a particular R statement
}

