% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check.chunk}
\alias{check.chunk}
\title{Check the student's solution of an exercise contained in stud.file}
\usage{
check.chunk(chunk.ind, ps = get.ps(), verbose = FALSE,
  stud.code = ps$cdt$stud.code[[chunk.ind]],
  stud.env = make.chunk.stud.env(chunk.ind, ps), expect.change = FALSE)
}
\arguments{
\item{ex.name}{The name of the exercise}

\item{stud.code}{The code of the student's solution as a string (or vector of strings)}
}
\description{
Check the student's solution of an exercise contained in stud.file
}

