\name{check.exercise}
\alias{check.exercise}
\title{Check the student's solution of an exercise contained in stud.file}
\usage{
\method{check}{exercise}(ex.name, stud.code, ps = get.ps(), verbose = FALSE)
}
\arguments{
  \item{ex.name}{The name of the exercise}

  \item{stud.code}{The code of the student's solution as a
  string (or vector of strings)}
}
\description{
Check the student's solution of an exercise contained in
stud.file
}

