% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_for_ps.r
\name{check.expr}
\alias{check.expr}
\title{Test: Compare the expression check.expr evaluated in the student's environment with the solution correct.expr}
\usage{
check.expr(check.expr, correct.expr,
  failure.message = "{{check_expr}} has the wrong values!",
  success.message = "Great, {{check_expr}} seems correct.", part = NULL,
  ps = get.ps(), stud.env = ps$stud.env, verbose = FALSE,
  unsubst.check.expr = NULL, unsubst.correct.expr = NULL,
  str.check.expr = NULL, str.correct.expr = NULL,
  tol = .Machine$double.eps^0.5)
}
\arguments{
\item{check.expr}{the expression to be checked}

\item{correct.expr}{the correct expression}

\item{vars}{a variable name or vector of variable names}

\item{exists}{shall existence be checked (similar length, class, values)}

\item{failure.exists}{a message that is shown if the variable does not exists (similar the other failure.??? variables)}

\item{failure.message.add}{a text that will be added to all failure messages}
}
\description{
Test: Compare the expression check.expr evaluated in the student's environment with the solution correct.expr
}

