% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check.file.exists}
\alias{check.file.exists}
\title{Check whether a given file exists}
\usage{
check.file.exists(file,
  failure.message = paste0("Sorry, but I cannot find the file \\"", file, "\\" in your current working directory."),
  success.message = paste0("Great, I have found the file \\"", file, "\\"!"),
  ps = get.ps(), part = NULL, ...)
}
\description{
Check whether a given file exists
}

