% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check.package}
\alias{check.package}
\title{Tests whether a package is loaded. Not yet implemented for speed reasons}
\usage{
check.package(package)
}
\description{
Tests whether a package is loaded. Not yet implemented for speed reasons
}

