\name{check.var}
\alias{check.var}
\title{Test: Compare students variables with those from the given solutions}
\usage{
  check.var(vars, exists = FALSE, length = FALSE,
    class = FALSE, values = FALSE, verbose = FALSE,
    student.env = get.student.env(),
    sol.env = get.sol.env())
}
\description{
  Test: Compare students variables with those from the
  given solutions
}

