% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{clean_cmd}
\alias{clean_cmd}
\title{Describe data frame}
\usage{
clean_cmd(cmd, selected_data, on_server = FALSE)
}
\arguments{
\item{cmd}{A string that stores the completion from GTP3.}

\item{selected_data, }{name of the selected dataset.}

\item{on_server, }{whether or not running on the server.}

\item{df}{a data frame}

\item{list_levels}{whether to list levels for factors}

\item{relevant_var}{a list of variables mentioned by the user}

\item{head}{whether to list the first few rows}
}
\value{
Returns a cleaned up version, so that it could be executed as R command.
Clean up R commands generated by GTP

The response from GTP3 sometimes contains strings that are not R commands.

Returns a cleaned up version, so that it could be executed as R command.
}
\description{
Returns information on data frame describing columns.
}
