% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{clean_cmd}
\alias{clean_cmd}
\title{Clean up R commands generated by GTP}
\usage{
clean_cmd(cmd, selected_data, on_server = FALSE)
}
\arguments{
\item{cmd}{A string that stores the completion from GTP3.}

\item{selected_data, }{name of the selected dataset.}

\item{on_server, }{whether or not running on the server.}
}
\value{
Returns a cleaned up version, so that it could be executed as R command.
}
\description{
The response from GTP3 sometimes contains strings that are not R commands.
}
