\name{create.ps}
\alias{create.ps}
\title{Generate a problem set from a solution file}
\usage{
create.ps(sol.file, ps.name, user.name = "ENTER A USER NAME HERE",
  sol.user.name = "Jane Doe", dir = getwd(), header = "", footer = "",
  libs = NULL)
}
\description{
Generates _struc.r file, .rps file, empty problem set .r
and .rmd files and a sample solution .rmd file (overwrites
existing files)
}

