% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create.ps}
\alias{create.ps}
\title{Generate a problem set from a solution file}
\usage{
create.ps(sol.file, ps.name = NULL, user.name = "ENTER A USER NAME HERE",
  sol.user.name = "Jane Doe", dir = getwd(), header = "", footer = "",
  libs = NULL, stop.when.finished = FALSE, extra.code.file = NULL,
  var.txt.file = NULL, rps.has.sol = TRUE)
}
\description{
Generates  _struc.r file, .rps file, empty problem set .r and .rmd files
and a sample solution .rmd file (overwrites existing files)
}

