\name{create.stud.ps}
\alias{create.stud.ps}
\title{Internal function to generate a problem set skeleton for a student and save it in a file}
\usage{
  create.stud.ps(ps, file = ps$stud.file,
    ps.dir = "C:/...")
}
\description{
  Internal function to generate a problem set skeleton for
  a student and save it in a file
}

