\name{create.stud.ps.rmd}
\alias{create.stud.ps.rmd}
\title{Internal function to generate a problem set skeleton for a student and save it in a file}
\usage{
create.stud.ps.rmd(ps, file = paste0(ps$stud.path, "/", ps$name, ".rmd"),
  ps.dir = "C:/...", user.name = "ENTER A USER NAME HERE", header = NULL,
  libs = NULL, footer = NULL)
}
\description{
Internal function to generate a problem set skeleton for a
student and save it in a file
}

