% Generated by roxygen2 (4.0.1): do not edit by hand
\name{data}
\alias{data}
\title{Overwrite the base function data, copy data by default into the calling environment instead of the global environment}
\usage{
data(..., envir = parent.frame())
}
\description{
Overwrite the base function data, copy data by default into the calling environment instead of the global environment
}

