% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{describe_df}
\alias{describe_df}
\title{Describe a data frame}
\usage{
describe_df(df, list_levels = FALSE, relevant_var = NULL, send_head = TRUE)
}
\arguments{
\item{df}{A data frame}

\item{list_levels}{Logical; whether to list levels for categorical variables}

\item{relevant_var}{A character vector of variable names to focus on (optional)}

\item{send_head}{Logical; whether to include a preview of the data frame}
}
\value{
A character string summarizing the data frame
}
\description{
Provides a description of a data frame, including numeric, character, and categorical variables.
}
