% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract.exercise.code}
\alias{extract.exercise.code}
\title{Extracts the stud's code of a given exercise}
\usage{
extract.exercise.code(ex.name, stud.code = ps$stud.code, ps = get.ps(),
  warn.if.missing = TRUE)
}
\description{
Extracts the stud's code of a given exercise
}

