% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awards.r
\name{give.award}
\alias{give.award}
\title{Used in a solution file. Give an award to a student who made it so far}
\usage{
give.award(award.name, award = ps$rps$awards[[award.name]],
  user = get.user(), ps = get.ps())
}
\description{
Used in a solution file. Give an award to a student who made it so far
}

