% Generated by roxygen2 (4.0.1): do not edit by hand
\name{give.award}
\alias{give.award}
\title{Used in a solution file. Give an award to a student who made it so far}
\usage{
give.award(award.name, award = ps$rps$awards[[award.name]],
  user = get.user(), ps = get.ps())
}
\description{
Used in a solution file. Give an award to a student who made it so far
}

