% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hint.for.assign}
\alias{hint.for.assign}
\title{Default hint for an assignment}
\usage{
hint.for.assign(expr, ps = get.ps(), env = ps$stud.env,
  stud.expr.li = ps$stud.expr.li, part = ps$part, s3.method = NULL,
  expr.object = NULL, ...)
}
\description{
Default hint for an assignment
}

