% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hint.for.compute}
\alias{hint.for.compute}
\title{Default hint for a compute block}
\usage{
hint.for.compute(expr, hints.txt = NULL, var = "", ps = get.ps(),
  env = ps$stud.env, stud.expr.li = ps$stud.expr.li, part = ps$part, ...)
}
\description{
Default hint for a compute block
}

