\name{holds.true}
\alias{holds.true}
\title{Test: Does a certain condition on the stud's generated variables hold true}
\usage{
holds.true(cond, short.message = failure.message,
  failure.message = "Failure in holds.true",
  success.message = "Great, the condition {{cond}} holds true in your solution!",
  hint.name = NULL, part = NULL, ex = get.ex(), stud.env = ex$stud.env,
  cond.str = NULL, ...)
}
\description{
Test: Does a certain condition on the stud's generated
variables hold true
}

