% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_for_ps.r
\name{hypothesis.test.result}
\alias{hypothesis.test.result}
\title{A helper function for hypothesis test about whether student solution is correct}
\usage{
hypothesis.test.result(p.value, alpha.warning = 0.05, alpha.failure = 1e-04,
  verbose = FALSE)
}
\description{
A helper function for hypothesis test about whether student solution is correct
}

