% Generated by roxygen2 (4.0.1): do not edit by hand
\name{init.ps}
\alias{init.ps}
\title{Initialize a problem set for the student}
\usage{
init.ps(ps.name, dir = getwd(), stud.short.file = paste0(ps.name, ".Rmd"),
  rps.file = paste0(dir, "/", ps.name, ".rps"), log.file = paste0(dir, "/",
  ps.name, ".log"))
}
\arguments{
\item{ps.name}{the name of the problem set}

\item{dir}{the path in which the stud has stored his file}

\item{stud.hort.file}{the file name (without path) of the .rmd problem set file}
}
\description{
Initialize a problem set for the student
}

