% Generated by roxygen2 (4.0.1): do not edit by hand
\name{name.rmd.chunks}
\alias{name.rmd.chunks}
\title{Set default names for the chunks of problem set rmd files}
\usage{
name.rmd.chunks(rmd.file = NULL, txt = readLines(rmd.file),
  only.empty.chunks = TRUE, keep.options = TRUE)
}
\description{
Set default names for the chunks of problem set rmd files
}

