% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{prep_input}
\alias{prep_input}
\title{Prepare User input.}
\usage{
prep_input(txt, selected_data, df, use_python)
}
\arguments{
\item{txt}{A string that stores the user input.}

\item{selected_data}{Name of the dataset.}

\item{df}{the data frame}

\item{use_python}{whether or not using python instead of R}
}
\value{
Returns a cleaned up version, so that it could be sent to GPT.
}
\description{
The response from GPT3 sometimes contains strings that are not R commands.
}
