% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_package.r
\name{rtutor.app.skel}
\alias{rtutor.app.skel}
\title{Generate a skeleton for a shinyapps.io app of a problem set}
\usage{
rtutor.app.skel(ps.name, app.name = ps.name, app.dir,
  rps.app = !is.null(rps.dir), pkg.name = NULL, rps.file = paste0(ps.name,
  ".rps"), rps.dir = NULL, overwrite = FALSE,
  github.user = "GITHUB_USERNAME", libs = NULL, ...)
}
\arguments{
\item{ps.name}{Name of the problem set}

\item{app.name}{Name of your app. Should have no white spaces or special characters}

\item{app.dir}{Your local directory to which you want to deploy your app files}

\item{rps.app}{locgical. If TRUE create an app based on an .rps file. Otherwise create the app based on a problem set package that is hosted on Github.}

\item{pkg.name}{If you create the app from a package this is the name of your package.}

\item{rps.file}{The name of your rps file without directory if you create the app from a .rps file}

\item{rps.dir}{the folder of your rps.file}
}
\description{
Generate a skeleton for a shinyapps.io app of a problem set
}

