% Generated by roxygen2 (4.0.1): do not edit by hand
\name{run.ps}
\alias{run.ps}
\title{Run a problem set from a package in the browser}
\usage{
run.ps(user.name, ps.name = info$ps[1], dir = getwd(),
  info = ps.pkg.info(), ...)
}
\description{
Only works if a package with problem sets is loaded.
For problem sets stored in a local .rps file use show.ps() instead
}

