% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{save_comments}
\alias{save_comments}
\title{Save user feedback}
\usage{
save_comments(date, time, comments, helpfulness, experience)
}
\arguments{
\item{date}{Date in the format of "2023-01-04"}

\item{time}{Time "13:05:12"}

\item{comments, }{user request}

\item{helpfulness}{rating}

\item{experience}{R experience}
}
\value{
nothing
}
\description{
Save user feedback
}
