% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{save_data}
\alias{save_data}
\title{Saves user queries, code, and error status}
\usage{
save_data(
  date,
  time,
  request,
  code,
  error_status,
  data_str,
  dataset,
  session,
  filename,
  filesize,
  chunk,
  api_time,
  tokens,
  language
)
}
\arguments{
\item{date}{Date in the format of "2023-01-04"}

\item{time}{Time "13:05:12"}

\item{request, }{user request}

\item{code}{AI generated code}

\item{filename}{name of the uploaded file}

\item{filesize}{size}

\item{chunk, }{id, from 1, 2, ...}

\item{api_time}{time in seconds for API response}

\item{tokens}{total completion tokens}

\item{error}{status, TRUE, error}
}
\value{
nothing
}
\description{
Saves user queries, code, and error status
}
