% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_rtutor.r
\name{show.ps}
\alias{show.ps}
\title{Run a problem set in the webbroser (or in the viewer pane).}
\usage{
show.ps(ps.name, user.name = "Seb", sav.file = NULL,
  load.sav = !is.null(sav.file), sample.solution = FALSE,
  run.solved = load.sav, import.rmd = FALSE, rmd.file = paste0(ps.name,
  "_", user.name, "_export.rmd"), launch.browser = TRUE,
  catch.errors = TRUE, dir = getwd(), rps.dir = dir,
  offline = !can.connect.to.MathJax(), left.margin = 2, right.margin = 2,
  is.solved, make.web.app = FALSE, make.session.ps = make.web.app,
  save.nothing = FALSE, show.solution.btn = TRUE,
  disable.graphics.dev = TRUE, clear.user = FALSE,
  check.whitelist = !is.null(wl), wl = NULL, verbose = FALSE,
  html.data.frame = TRUE, table.max.rows = 25, round.digits = 8,
  signif.digits = 8, knit.print.opts = make.knit.print.opts(html.data.frame
  = html.data.frame, table.max.rows = table.max.rows, round.digits =
  round.digits, signif.digits = signif.digits), precomp = FALSE,
  noeval = FALSE, ...)
}
\arguments{
\item{load.sav}{shall the last saved be loaded?}

\item{sample.solution}{shall the sample solution be shown}

\item{run.solved}{if sample.solution or load.sav shall the correct chunks be automatically run when the problem set is loaded? (Starting the problem set then may take quite a while)}

\item{import.rmd}{shall the solution be imported from the rmd file specificed in the argument rmd.file}

\item{catch.errors}{by default TRUE only set FALSE for debugging purposes in order to get a more informative traceback()}

\item{offline}{(FALSE or TRUE) Do you have no internet connection. By default it is checked whether RTutor can connect to the MathJax server. If you have no internet connection, you cannot render mathematic formulas. If RTutor wrongly thinks you have an internet connection, while you don't, your chunks may not show at all. If you encounter this problem, set manually offline=TRUE.}

\item{is.solved}{DEPRECEATED}

\item{html.data.frame}{shall data.frames and matrices be printed as html table if a chunk is checked? (Default=TRUE)}

\item{table.max.rows}{the maximum number of rows that is shown if a data.frame is printed as html.table}

\item{round.digits}{the number of digits that printed data.frames shall be rounded to}

\item{lauch.browser}{if TRUE (default) show the problem set in the browser. Otherwise it is shown in the RStudio viewer pane}
}
\description{
... contains parameters specified in init.shiny.ps. They are explained here.
}

