% Generated by roxygen2 (4.0.1): do not edit by hand
\name{show.success.message}
\alias{show.success.message}
\title{Simply shows a success message when this test is reached for the first time!}
\usage{
show.success.message(success.message, ...)
}
\description{
Simply shows a success message when this test is reached for the first time!
}

