% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sigma.test}
\alias{sigma.test}
\title{Test for variance (copied from TeachingDemos)}
\usage{
sigma.test(x, sigma = 1, sigmasq = sigma^2, alternative = c("two.sided",
  "less", "greater"), conf.level = 0.95, ...)
}
\description{
Test for variance (copied from TeachingDemos)
}

