\name{test.mean}
\alias{test.mean}
\title{Test: The mean of the distribution from which a vector of random numbers has been drawn}
\usage{
  test.mean(vec, true.val, test = "t.test", short.message,
    warning.message, failure.message, ex = get.ex(),
    stud.env = ex$stud.env, ...)
}
\description{
  Test: The mean of the distribution from which a vector of
  random numbers has been drawn
}

