\name{test.normality}
\alias{test.normality}
\title{Test: Has a vector of random numbers been drawn from a normal distribution?}
\usage{
  test.normality(vec, alpha.warning = 0.05,
    alpha.fail = 1e-04, student.env = get.student.env(),
    ..., digits = 7, verbose = FALSE)
}
\description{
  Test: Has a vector of random numbers been drawn from a
  normal distribution?
}

