% Generated by roxygen2 (4.0.1): do not edit by hand
\name{test.normality}
\alias{test.normality}
\title{Test: Has a vector of random numbers been drawn from a normal distribution?}
\usage{
test.normality(vec, short.message, warning.message, failure.message,
  ps = get.ps(), stud.env = ps$stud.env,
  success.message = "Great, I cannot statistically reject that {{var}} is indeed normally distributed!",
  part = NULL, ...)
}
\description{
Test: Has a vector of random numbers been drawn from a normal distribution?
}

