\name{test.variance}
\alias{test.variance}
\title{Test: The variance of the distribution from which a vector of random numbers has been drawn}
\usage{
  test.variance(vec, true.val, test = "t.test",
    short.message, warning.message, failure.message,
    success.message = "Great, I cannot statistically reject that {{var}} has the desired variance {{vari_sol}}!",
    ex = get.ex(), stud.env = ex$stud.env, ...)
}
\description{
  Test: The variance of the distribution from which a
  vector of random numbers has been drawn
}

