% Generated by roxygen2 (4.0.1): do not edit by hand
\name{test.variance}
\alias{test.variance}
\title{Test: The variance of the distribution from which a vector of random numbers has been drawn}
\usage{
test.variance(vec, true.val, test = "t.test", short.message, warning.message,
  failure.message,
  success.message = "Great, I cannot statistically reject that {{var}} has the desired variance {{vari_sol}}!",
  ps = get.ps(), stud.env = ps$stud.env, part = NULL, ...)
}
\description{
Test: The variance of the distribution from which a vector of random numbers has been drawn
}

