% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTutor.R
\name{zip.solution}
\alias{zip.solution}
\title{Create a zip file of your solution that can be submitted}
\usage{
zip.solution(ps = get.ps(), user.name = get.user()$name,
  dir = ps$stud.path, ask = TRUE)
}
\description{
Only works after you have once checked your problem set!
}

